separate the tests of the path-parameter example
