+++ examples/path-parameter/src/main.rs
@@ -34,44 +34,4 @@ async fn user_page(Path(name): Path<String>) -> Html<String> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::{body::Body, http::Request, http::StatusCode};
-    use http_body_util::BodyExt;
-    use tower::ServiceExt;
-
-    #[tokio::test]
-    async fn test_main_page() {
-        let response = app()
-            .oneshot(Request::builder().uri("/").body(Body::empty()).unwrap())
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response.into_body();
-        let bytes = body.collect().await.unwrap().to_bytes();
-        let html = String::from_utf8(bytes.to_vec()).unwrap();
-
-        assert!(html.contains(r#"<a href="/user/foo">/user/foo</a><br>"#));
-    }
-
-    #[tokio::test]
-    async fn test_user_page() {
-        let response = app()
-            .oneshot(
-                Request::builder()
-                    .uri("/user/qqrq")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response.into_body();
-        let bytes = body.collect().await.unwrap().to_bytes();
-        let html = String::from_utf8(bytes.to_vec()).unwrap();
-
-        assert_eq!(html, "Hello, qqrq!");
-    }
-}
+mod tests;
